feat: remove hard deps on s3 - removes S3 config loading from yace init, instead accepting function for loading config file to allow for different storage methods or hardcoding for tests
