+++ pkg/types/types.go
@@ -13,7 +13,7 @@ type Logger interface {
 }
 
 type MetricGatherer interface {
-	Init() error
+	Init(func() ([]byte, error)) error
 	CollectMetrics(Logger, Config) error
 	ExtractMetrics(Logger) ([]*io_prometheus_client.MetricFamily, error)
 	GetRegistry() *prometheus.Registry
